Add Cache-control headers to Javascript & CSS
Every time a page is (re)loaded, the ESP8266 has to send the stylesheet and javascript before the page loads. Although this is minimal (20ms in dev tools Networks pane), it can improve page load speeds when the client device is spamming network connectivity check requests.

Adding the `Cache-Control` headers makes the browser store these files in memory (for 300 seconds, could be increased). The same could be done for the HTML files, but that would allow the pages to load even when not connected to the ESP8266.