No more static stuff
Moved the print functions to the function.h and included them in language using extern so not every class has to "extern" them again.
Also fixed some type comparisons and made removed the static from the progmem variables.
C code is super messy! But I'm not going to rewrite it again so deal with it :D