Station print buffer
Don't print new scan results out immediately, but set a flag and print them every second. That way, the sniffer callback is as fast as possible, not blocking anything. 
Before it was possible that when a new station was discovered while the output from the last was not fully printed, the device could crash.